+++ src/main/java/seedu/duke/transaction/TransactionList.java
@@ -3,6 +3,7 @@ package seedu.duke.transaction;
 import seedu.duke.exception.InvalidTransactionException;
 import seedu.duke.exception.TransactionNotFoundException;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import static seedu.duke.exception.message.ExceptionMessages.MESSAGE_ITEM_TRANSACTION_OVERLAP;
@@ -65,7 +66,7 @@ public class TransactionList {
      */
 
     //@@author jorellesee
-    public Transaction updateTransaction(String transactionId, int duration, double moneyTransacted)
+    public Transaction updateTransaction(String transactionId, int duration, BigDecimal moneyTransacted)
             throws TransactionNotFoundException, InvalidTransactionException {
         for (int i = 0; i < this.transactionList.size(); ++i) {
             Transaction tx = this.transactionList.get(i);
@@ -123,7 +124,7 @@ public class TransactionList {
     public double getTotalMoneyTransacted() {
         int totalProfit = 0;
         for (Transaction transaction : transactionList) {
-            totalProfit += transaction.getMoneyTransacted();
+            totalProfit += transaction.getMoneyTransacted().doubleValue();
         }
         return totalProfit;
     }
